Add break line on message input
